requirements.txt
PyYAML==6.0.2
feedparser==6.0.11
requests==2.32.3
beautifulsoup4==4.12.3
